[API,INFRA] move type_list.hpp to own subfolder
