+++ include/seqan3/core/tuple_utility.hpp
@@ -18,7 +18,7 @@
 
 #include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/pod_tuple.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/core/type_traits/basic.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
