+++ include/seqan3/core/type_list/traits.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 
 // ----------------------------------------------------------------------------
 // seqan3::pack_traits::detail
