+++ include/seqan3/core/type_list/type_list.hpp
@@ -25,7 +25,7 @@ namespace seqan3
 
 /*!\brief Type that contains multiple types, an alias for
  * [meta::list](https://ericniebler.github.io/range-v3/structmeta_1_1list.html).
- * \ingroup core
+ * \ingroup type_list
  */
 template <typename ... types>
 using type_list = meta::list<types...>;
@@ -40,7 +40,7 @@ namespace seqan3::detail
 // ----------------------------------------------------------------------------
 
 /*!\brief Auxiliary concept that checks whether a type is a specialisation of seqan3::type_list.
- * \ingroup core
+ * \ingroup type_list
  */
 template <typename t>
 SEQAN3_CONCEPT TypeList = is_type_specialisation_of_v<t, type_list>;
