+++ include/seqan3/io/alignment_file/input_format_concept.hpp
@@ -21,7 +21,7 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/quality/aliases.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/io/alignment_file/header.hpp>
 #include <seqan3/io/alignment_file/input_options.hpp>
 #include <seqan3/io/alignment_file/sam_tag_dictionary.hpp>
