+++ include/seqan3/io/sequence_file/input_format_concept.hpp
@@ -20,7 +20,7 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/quality/aliases.hpp>
 #include <seqan3/alphabet/quality/phred42.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/io/sequence_file/input_options.hpp>
 
 namespace seqan3::detail
