+++ include/seqan3/io/structure_file/input_format_concept.hpp
@@ -21,7 +21,7 @@
 #include <seqan3/alphabet/nucleotide/rna5.hpp>
 #include <seqan3/alphabet/structure/structured_rna.hpp>
 #include <seqan3/alphabet/structure/wuss.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/io/structure_file/input_options.hpp>
 
 namespace seqan3::detail
