+++ test/snippet/core/algorithm/for_each_type_list.cpp
@@ -2,7 +2,7 @@
 #include <string>
 
 #include <seqan3/core/algorithm/parameter_pack.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/core/debug_stream.hpp>
 
 using namespace seqan3;
