+++ test/unit/alignment/scoring_scheme_test.cpp
@@ -17,7 +17,7 @@
 #include <seqan3/alphabet/aminoacid/all.hpp>
 #include <seqan3/alphabet/nucleotide/all.hpp>
 #include <seqan3/alphabet/quality/all.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/test/cereal.hpp>
 
 using namespace seqan3;
