+++ test/unit/alphabet/nucleotide/nucleotide_conversion_integration_test.cpp
@@ -10,7 +10,7 @@
 #include <meta/meta.hpp>
 
 #include <seqan3/alphabet/nucleotide/all.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 
 using namespace seqan3;
