+++ test/unit/core/algorithm/parameter_pack_test.cpp
@@ -9,7 +9,7 @@
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/core/algorithm/parameter_pack.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 
 using namespace seqan3;
