+++ test/unit/core/simd/detail/builtin_simd_test.cpp
@@ -10,7 +10,7 @@
 #include <seqan3/core/algorithm/parameter_pack.hpp>
 #include <seqan3/core/simd/concept.hpp>
 #include <seqan3/core/simd/detail/builtin_simd.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 #include <seqan3/std/type_traits>
 
 #include <iostream>
