+++ test/unit/core/type_traits/template_inspection_test.cpp
@@ -11,7 +11,7 @@
 
 #include <seqan3/core/type_traits/concept.hpp>
 #include <seqan3/core/type_traits/template_inspection.hpp>
-#include <seqan3/core/type_list.hpp>
+#include <seqan3/core/type_list/type_list.hpp>
 
 using namespace seqan3;
